app removed start
