+++ include/moss/core/ecs.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
 
-#include <moss/core/ecs/primitives.hpp>
-#include <moss/core/ecs/read.hpp>
+#include <moss/core/ecs/component.hpp>
+#include <moss/core/ecs/system.hpp>
+#include <moss/core/ecs/package.hpp>
