--- include/moss/core/ecs/primitives/primitives.hpp
@@ -1,20 +1,0 @@
-/*
-core/ecs/primitives.hpp
-
-Defines the standard/base component,
-package and system classes/structs.
-
-*/
-
-#pragma once
-
-
-#include <moss/meta/libs.hpp>
-#include <moss/core/contex.hpp>
-
-
-namespace moss {
-
-
-
-} // moss
